Move tests into new folder
